Update ApiResponse ERROR from exceptions
