+++ src/main/java/org/tdl/vireo/controller/advice/CustomResponseEntityExceptionHandler.java
@@ -5,8 +5,6 @@ import static java.lang.String.format;
 
 import javax.persistence.EntityNotFoundException;
 
-import java.io.PrintWriter;
-import java.io.StringWriter;
 import java.nio.file.NoSuchFileException;
 
 import org.hibernate.exception.ConstraintViolationException;
@@ -15,7 +13,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
-import org.springframework.lang.NonNull;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -72,7 +69,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(CONSTRAINT_VIOLATION_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(DataIntegrityViolationException.class)
@@ -81,7 +78,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(DATA_INTEGRITY_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(EntityNotFoundException.class)
@@ -90,7 +87,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(ENTITY_NOT_FOUND_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(NoSuchFileException.class)
@@ -99,7 +96,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(FILE_NOT_FOUND_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(MultipartException.class)
@@ -108,7 +105,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(PAYLOAD_TOO_LARGE_TEMPLATE, maxFileSize);
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositBadGatewayException.class)
@@ -117,7 +114,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_BAD_GATEWAY_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositBadRequestException.class)
@@ -126,7 +123,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_BAD_REQUEST_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositConflictException.class)
@@ -135,7 +132,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_CONFLICT_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositForbiddenException.class)
@@ -144,7 +141,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_FORBIDDEN_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositGatewayTimeoutException.class)
@@ -153,7 +150,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_GATEWAY_TIMEOUT_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositInternalServerErrorException.class)
@@ -162,7 +159,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_INTERNAL_ERROR_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositNotFoundException.class)
@@ -171,7 +168,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_NOT_FOUND_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositNotImplementedException.class)
@@ -180,7 +177,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_NOT_IMPLEMENTED_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositRequestTimeoutException.class)
@@ -189,7 +186,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_REQUEST_TIMEOUT_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositServiceUnavailableException.class)
@@ -198,7 +195,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_SERVICE_UNAVAILABLE_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositUnauthorizedException.class)
@@ -207,7 +204,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_UNAUTHORIZED_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
     @ExceptionHandler(SwordDepositUnprocessableEntityException.class)
@@ -216,16 +213,7 @@ public class CustomResponseEntityExceptionHandler extends ResponseEntityExceptio
         String message = format(SWORD_UNPROCESSABLE_TEMPLATE, exception.getMessage());
         logger.error(message);
         logger.debug(message, exception);
-        return ApiResponse.fromError(ERROR, message, serializeStacktrace(exception));
-    }
-
-    private static String serializeStacktrace(@NonNull Exception exception) {
-        // Convert stack trace to string
-        StringWriter stringWriter = new StringWriter();
-        PrintWriter printWriter = new PrintWriter(stringWriter);
-        exception.printStackTrace(printWriter);
-
-        return stringWriter.toString();
+        return ApiResponse.fromException(ERROR, message, exception);
     }
 
 }
